careercompass/backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
google-generativeai==0.7.2
pyresparser==1.0.6
nltk==3.8.1
python-docx==0.8.12